refactor: split session.rs into connection, transfer, and navigation modules
